% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big5}
\alias{big5}
\title{Example Big-5 personality data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100000 rows and 25 columns.
}
\source{
\url{https://www.kaggle.com/datasets/tunguz/big-five-personality-test}
}
\usage{
big5
}
\description{
This data was pre-processed for this package specifically.
}
\seealso{
\code{\link[=synp_get_param]{synp_get_param()}}
}
\keyword{datasets}
