% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Birthweight}
\alias{Birthweight}
\title{Example Birthweight data}
\format{
An object of class \code{data.frame} with 47 rows and 16 columns.
}
\source{
\url{https://www.sheffield.ac.uk/mash/statistics/datasets}
}
\usage{
Birthweight
}
\description{
This dataset contains information on new born babies and their parents.
}
\details{
\describe{
\item{ID}{Baby number}
\item{length}{Length of baby (cm)}
\item{Birthweight}{Weight of baby (kg)}
\item{headcirumference}{Head Circumference}
\item{Gestation}{Gestation (weeks)}
\item{smoker}{Mother smokes 1 = smoker 0 = non-smoker}
\item{motherage}{Maternal age}
\item{mnocig}{Number of cigarettes smoked per day by mother}
\item{mheight}{Mothers height (cm)}
\item{mppwt}{Mothers pre-pregnancy weight (kg)}
\item{fage}{Father's age}
\item{fedyrs}{Father’s years in education}
\item{fnocig}{Number of cigarettes smoked per day by father}
\item{fweight}{Father's weight (kg)}
\item{lowbwt}{Low birth weight, 0 = No and 1 = yes (cut off 2.6kg)}
\item{mage35}{Mother over 35, 0 = No and 1 = yes}
}
}
\keyword{datasets}
