% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{synp_gen_syndat}
\alias{synp_gen_syndat}
\title{Generate synthetic data from a parameter list}
\usage{
synp_gen_syndat(par_list, n = 1000)
}
\arguments{
\item{par_list}{the parameter list}

\item{n}{sample size (default = 1000)}
}
\value{
a data frame
}
\description{
Generate synthetic data from a parameter list
}
