% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{synp_get_param}
\alias{synp_get_param}
\title{Extract parameter list from synds object}
\usage{
synp_get_param(df, synds)
}
\arguments{
\item{df}{the original data frame}

\item{synds}{the fitted synds object}
}
\value{
a list of data frames
}
\description{
Extract parameter list from synds object
}
\details{
The synds object should be run with a parametric method and with the argument \code{models = TRUE}
}
\examples{

# small example with big5 data
synds <- syn(big5, method="norm", models=TRUE)

# get model parameters
model_par <- synp_get_param(big5, synds)

# show model parameters
model_par

}
\seealso{
\code{\link[=synp_write_sheets]{synp_write_sheets()}}, \code{\link[=synp_read_sheets]{synp_read_sheets()}}, \code{\link[=synp_gen_syndat]{synp_gen_syndat()}}
}
