% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{synp_read_sheets}
\alias{synp_read_sheets}
\title{Read all the sheets in an xlsx workbook into a list of data frames}
\usage{
synp_read_sheets(path, ...)
}
\arguments{
\item{path}{the file path of the xslx workbook}

\item{...}{arguments passed to \code{readxl::read_xlsx()}}
}
\value{
a list of data frames
}
\description{
Read all the sheets in an xlsx workbook into a list of data frames
}
