% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{synp_write_sheets}
\alias{synp_write_sheets}
\title{Write the parameter list to an excel file for exporting}
\usage{
synp_write_sheets(par_list, path)
}
\arguments{
\item{par_list}{the parameter list}

\item{path}{the file path of the xslx workbook}
}
\description{
Write the parameter list to an excel file for exporting
}
